python-scripts/requirements.txt
# TriPoint API dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0
python-dotenv>=1.0
requests>=2.28
WazeRouteCalculator>=0.14
aiosqlite>=0.19.0
stripe>=7.0
itsdangerous>=2.1
reportlab>=4.0
